# Optimizations to handle large workflows