--- a/snakemake/dag.py
+++ b/snakemake/dag.py
@@ -9,7 +9,7 @@ import shutil
 import textwrap
 import time
 import tarfile
-from collections import defaultdict, Counter
+from collections import defaultdict, Counter, deque
 from itertools import chain, filterfalse, groupby
 from functools import partial
 from pathlib import Path
@@ -937,6 +937,7 @@ class DAG:
 
         queue = list(filter(reason, candidates))
         visited = set(queue)
+        candidates_set = set(candidates)
         while queue:
             job = queue.pop(0)
             _needrun.add(job)
@@ -949,7 +950,7 @@ class DAG:
                     queue.append(job_)
 
             for job_, files in depending[job].items():
-                if job_ in candidates:
+                if job_ in candidates_set:
                     reason(job_).updated_input_run.update(files)
                     if not job_ in visited:
                         visited.add(job_)
@@ -1070,9 +1071,11 @@ class DAG:
         if jobs is None:
             jobs = self.needrun_jobs
 
+        potential_new_ready_jobs = False
         candidate_groups = set()
         for job in jobs:
             if not self.finished(job) and self._ready(job):
+                potential_new_ready_jobs = True
                 if job.group is None:
                     self._ready_jobs.add(job)
                 else:
@@ -1085,6 +1088,7 @@ class DAG:
             for group in candidate_groups
             if all(self._ready(job) for job in group)
         )
+        return potential_new_ready_jobs
 
     def get_jobs_or_groups(self):
         visited_groups = set()
@@ -1252,7 +1256,7 @@ class DAG:
 
         # mark depending jobs as ready
         # skip jobs that are marked as until jobs
-        self.update_ready(
+        potential_new_ready_jobs = self.update_ready(
             j
             for job in jobs
             for j in self.depending[job]
@@ -1281,6 +1285,9 @@ class DAG:
                 self.pull_container_imgs()
             if self.workflow.use_conda:
                 self.create_conda_envs()
+            potential_new_ready_jobs = True
+
+        return potential_new_ready_jobs
 
     def new_job(self, rule, targetfile=None, format_wildcards=None):
         """Create new job for given rule and (optional) targetfile.
@@ -1459,10 +1466,10 @@ class DAG:
 
     def bfs(self, direction, *jobs, stop=lambda job: False):
         """Perform a breadth-first traversal of the DAG."""
-        queue = list(jobs)
+        queue = deque(jobs)
         visited = set(queue)
         while queue:
-            job = queue.pop(0)
+            job = queue.popleft()
             if stop(job):
                 # stop criterion reached for this node
                 continue