Large workflows (njobs>10k) can become very slow due to some code which scales O(n^2). 

    * Change candidates in update_needrun from list to set 
       (enable quick member check)

    * Use deque for faster bfs jobs traversal
       (prevents memory copy due to popping of first item)

    * Optimize open_jobs using set operations.

    * Only run open_jobs if new jobs have potentially been added to ready jobs

    * Add short-circuits to the periodicity detector to improve its speed.