@johanneskoester 2020-10-15
Do you have an example number for how much speed gain to expect for large workflows?