@mhulsman 2020-10-16
It can become quite extreme. I tested a 283k jobs workflow.   Doing a dry run in a empty dir:
Old: 42248 seconds   (41118 seconds in update_needrun)
New: 668 seconds   (114 seconds in update_needrun). 

- is_periodic went from 54 seconds to 24 seconds. 

The open_jobs optimization was not benchmarked here, as this requires a non-dry run.